requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0
PyMuPDF==1.23.5
qrcode[pil]==7.4.2
Pillow==10.0.1
weasyprint
Werkzeug==2.3.7
faker
qrcode[pil]
PyYAML
